[SD-222] Move `Hashable` instance out of `impl`
