+++ snowdrop-core/src/Snowdrop/Core/Transaction.hs
@@ -37,6 +37,8 @@ type family TxRawImpl (txtype :: k) :: *
 
 newtype TxRaw txtype = TxRaw { unTxRaw :: TxRawImpl txtype }
 
+deriving instance Hashable (TxRawImpl txtype) => Hashable (TxRaw txtype)
+
 -- | Transaction which modifies state.
 -- There is also RawTx, which is posted on the blockchain.
 -- Ideally, RawStateTx and any action which modifies a state can be converted into StateStateTx.
@@ -45,6 +47,7 @@ data StateTx (txtype :: *) = StateTx
     , txBody  :: HChangeSet (TxComponents txtype)
     } deriving (Generic)
 
+instance (Hashable (HChangeSet (TxComponents txtype)), Hashable (TxProof txtype)) => Hashable (StateTx txtype)
 deriving instance (Eq (HChangeSet (TxComponents txtype)), Eq (TxProof txtype)) => Eq (StateTx txtype)
 deriving instance (Ord (HChangeSet (TxComponents txtype)), Ord (TxProof txtype)) => Ord (StateTx txtype)
 deriving instance (Show (HChangeSet (TxComponents txtype)), Show (TxProof txtype)) => Show (StateTx txtype)
